requirements.txt
# requirements.txt - Список зависимостей приложения

# Основные библиотеки
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.22.0
matplotlib>=3.5.0
plotly>=5.13.0

# Работа с API и данными
requests>=2.28.0
statsmodels>=0.13.5

# AI и интеграция с OpenAI
openai>=0.27.0

# Создание PDF
fpdf2>=2.7.0

# Управление многопоточностью
# (встроена в Python, но явно указываем, что используется)
# threading

# Для обработки очередей
# (встроена в Python, но явно указываем, что используется)
# queue

# Полезные утилиты
python-dotenv>=1.0.0